requirements.txt
requests
termcolor
discord.py==1.5.0
jishaku